+++ android/src/main/kotlin/top/kikt/imagescanner/core/utils/AndroidQDBUtils.kt
@@ -85,6 +85,11 @@ object AndroidQDBUtils : IDBUtils {
       val count = countMap[id]!!
 
       val entity = GalleryEntity(id, name, count, requestType, false)
+
+      if (option.containsPathModified) {
+        injectModifiedDate(context, entity)
+      }
+
       list.add(entity)
     }