+++ lib/src/entity.dart
@@ -42,6 +42,11 @@ class AssetPathEntity {
 
   late FilterOptionGroup filterOption;
 
+  /// The modification time of the latest asset contained in an album.
+  ///
+  /// See [FilterOptionGroup.containsPathModified]
+  DateTime? lastModified;
+
   /// The value used internally by the user.
   /// Used to indicate the value that should be available inside the path.
   RequestType get type => _type;
@@ -91,6 +96,7 @@ class AssetPathEntity {
       this.isAll = result.isAll;
       this.type = result.type;
       this.filterOption = filterOption;
+      this.lastModified = result.lastModified;
     }
   }