+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 - [CHANGELOG](#changelog)
-  - [1.0.7](#107)
+  - [1.1.0](#110)
   - [1.0.6](#106)
   - [1.0.4](#104)
   - [1.0.3](#103)
@@ -52,7 +52,9 @@
   - [0.0.2 update readme](#002-update-readme)
   - [0.0.1](#001)
 
-## 1.0.7
+## 1.1.0
+
+- Feature: `modified` of `AssetPathEntity`.
 
 - Fix: Order option of the `FilterOptionGroup`.